+++ tests/djapp/models.py
@@ -32,6 +32,11 @@ class MultifieldModel(models.Model):
     text = models.CharField(max_length=20)
 
 
+class MultifieldUniqueModel(models.Model):
+    slug = models.SlugField(max_length=20, unique=True)
+    text = models.CharField(max_length=20, unique=True)
+
+
 class AbstractBase(models.Model):
     foo = models.CharField(max_length=20)